requirements.txt
torch
torchvision
transformers
datasets
timm
evaluate
accelerate
pillow
streamlit
streamlit-drawable-canvas